+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0 (2018-06-26)
+
+- [added] Customizable logging middleware
+- [added] Configurable SAML support
+- [added] Customizable SAML handler
+
 ## 1.4.1 (2017-03-23)
 
 - [bugfix] Update dependency on phobos_db_checkpoint to support newer api (events without retry)